backend/requirements.txt
astropy==6.1.0
coverage==7.5.1
Django==5.0.6
django-cors-headers==4.3.1
django-filebrowser-no-grappelli==4.0.2
django-filter==24.2
django-oauth-toolkit==2.3.0
django-shibboleth-remoteuser==0.12
djangorestframework==3.15.1
drf-social-oauth2==2.2.0
drf-spectacular==0.27.2
gunicorn==22.0.0
h5py==3.11.0
numpy==1.26.4
pandas==2.2.2
psycopg2-binary==2.9.9
pyarrow==16.0.0
PyJWT==2.8.0
pytest==8.2.0
pytest-cov==5.0.0
pytest-django==4.8.0
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
tables-io==0.9.6
tables==3.9.2
uWSGI==2.0.25.1
pydantic==2.7.1
